+++ airflow/contrib/operators/dingding_operator.py
@@ -16,14 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.dindding.operators.dingding`."""
+"""This module is deprecated. Please use `airflow.providers.dinding.operators.dingding`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.providers.dindding.operators.dingding import DingdingOperator  # noqa
+from airflow.providers.dinding.operators.dingding import DingdingOperator  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.dindding.operators.dingding`.",
+    "This module is deprecated. Please use `airflow.providers.dinding.operators.dingding`.",
     DeprecationWarning, stacklevel=2
 )
