+++ docs/conf.py
@@ -242,7 +242,7 @@ exclude_patterns = [
     '_api/airflow/providers/cncf/kubernetes/index.rst',
     '_api/airflow/providers/databricks/index.rst',
     '_api/airflow/providers/datadog/index.rst',
-    '_api/airflow/providers/ddiscord/index.rst',
+    '_api/airflow/providers/discord/index.rst',
     '_api/airflow/providers/dindding/index.rst',
     '_api/airflow/providers/dingding/index.rst',
     '_api/airflow/providers/discord/index.rst',
